#22 move required methods to sentinel1image, add utils, rename methods
