+++ s1denoise/__init__.py
@@ -1,2 +1,2 @@
-from s1denoise.S1_TOPS_GRD_NoiseCorrection import Sentinel1Image
+from s1denoise.sentinel1image import Sentinel1Image
 __all__ = ['Sentinel1Image']
