--- s1denoise/scripts/s1_thermal_denoise.py
@@ -1,29 +1,0 @@
-#!/usr/bin/env python
-import sys
-import argparse
-
-from s1denoise import Sentinel1Image
-from s1denoise.utils import run_denoising
-
-def parse_args(args):
-    ''' parse input arguments '''
-    parser = argparse.ArgumentParser(
-            description="Remove thermal noise from Sentinel-1 TOPSAR EW GRDM")
-    parser.add_argument('ifile', type=str, help='input Sentinel-1 file in SAFE format')
-    parser.add_argument('ofile', type=str, help='output GeoTIFF file')
-    parser.add_argument('-db', '--decibel', action='store_true', help='Export in decibels')
-    parser.add_argument('-nn', '--no-negatives', action='store_true',
-                        help='Replace negative pixels with nearest smallest positive pixel')
-    parser.add_argument('-p', '--polarisation', default='HV',
-                        choices=['HH', 'HV', 'HHHV',],
-                        help='Process bands in these polarizations')
-    return parser.parse_args(args)
-
-
-if __name__ == '__main__':
-    args = parse_args(sys.argv[1:])
-    pols = {'HH': ['HH'],
-            'HV': ['HV'],
-            'HHHV': ['HH', 'HV']}[args.polarisation]
-    s1 = run_denoising(args.ifile, pols=pols, db=args.decibel, filter_negative=args.no_negatives)
-    s1.export(args.ofile, driver='GTiff')
